+++ HoneyDrunk.Data/HoneyDrunk.Data.Abstractions/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to HoneyDrunk.Data.Abstractions will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-04-25
+
+### Changed
+
+- Coordinated package train bump for the ADR-0005/0006 Data rollout.
+
 ## [0.2.0] - 2026-01-06
 
 ### Changed