+++ HoneyDrunk.Data/HoneyDrunk.Data.EntityFramework/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to HoneyDrunk.Data.EntityFramework will be documented in thi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-04-25
+
+### Changed
+
+- Coordinated package train bump for the ADR-0005/0006 Data rollout.
+
 ## [0.3.0] - 2026-02-15
 
 ### Added